[fix:35] adding masks to utils
