+++ src/Utils/validators.js
@@ -80,3 +80,45 @@ export const mascaraCEP = (cep) => {
   }
   return formattedCEP.replace(/(\d{5})(\d)/, "$1-$2");
 };
+
+export const handleCpfCnpjInput = (value) => {
+  const numericValue = value.replace(/\D/g, "");
+  if (numericValue.length <= 11) {
+    return numericValue
+      .replace(/(\d{3})(\d)/, "$1.$2")
+      .replace(/(\d{3})(\d)/, "$1.$2")
+      .replace(/(\d{3})(\d{1,2})$/, "$1-$2")
+      .slice(0, 14); // CPF formatado
+  } else {
+    return numericValue
+      .replace(/^(\d{2})(\d)/, "$1.$2")
+      .replace(/^(\d{2})\.(\d{3})(\d)/, "$1.$2.$3")
+      .replace(/\.(\d{3})(\d)/, ".$1/$2")
+      .replace(/(\d{4})(\d{1,2})$/, "$1-$2")
+      .slice(0, 18); // CNPJ formatado
+  }
+};
+
+export const isValidTelefone = (telefone) => {
+  const cleanedNumber = telefone.replace(/\D/g, ""); // Remove caracteres não numéricos
+
+  if (!cleanedNumber) {
+    return { isValid: true };
+  }
+
+  if (cleanedNumber.length != 10) {
+    return { isValid: false, message: "O telefone fornecido não é válido." };
+  }
+
+  return { isValid: true };
+};
+
+export const mascaraCelular = (celular) => {
+  let formattedCelular = celular.replace(/\D/g, "");
+  if (formattedCelular.length > 11) {
+    formattedCelular = formattedCelular.slice(0, 11);
+  }
+  return formattedCelular
+    .replace(/^(\d{2})(\d)/g, "($1) $2")
+    .replace(/(\d{5})(\d{4})$/, "$1-$2");
+};
